karin: overlay: Enable config_powerDecoupleInteractiveModeFromDisplay
Power Management: Specifies whether to decouple the interactive state of the
device from the display on/off state.

When false, setInteractive(..., true) will be called before the display is turned on
and setInteractive(..., false) will be called after the display is turned off.
This mode provides best compatibility for devices that expect the interactive
state to be tied to the display state.

When true, setInteractive(...) will be called independently of whether the display
is being turned on or off.  This mode enables the power manager to reduce
clocks and disable the touch controller while the display is on.

This resource should be set to "true" when a doze component has been specified
to maximize power savings but not all devices support it.

Refer to power.h for details.

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: Ie3d35901881183eb8f5e12973e9888358e4f73ea